backend/requirements-flask.txt
Flask==2.3.3
flask-cors==3.0.10
pillow==10.1.0
# adjust versions as needed